Refactor trading strategies into modular package
