+++ KryptoLowca/bot/trading_strategies.py
@@ -1,2259 +1,5 @@
-# trading_strategies_enhanced.py
-# -*- coding: utf-8 -*-
-
-"""
-Production-grade Trading Engine with Clean Architecture - Enhanced Version 2.0
-
-Key improvements for 10/10 rating:
-- Fixed all deprecated pandas methods
-- Implemented comprehensive risk management with stop-loss/take-profit
-- Added vectorized operations throughout
-- Comprehensive error handling with custom exceptions
-- Advanced performance metrics including Sortino, Omega ratios
-- Complete test coverage
-- Configuration management
-- Performance optimization with caching
-- Advanced logging and monitoring
-- Documentation following Google style
-- Type safety with strict mypy compliance
-"""
+"""Compatibility shim for the trading strategies package used by the bot."""
 
 from __future__ import annotations
 
-import hashlib
-import logging
-import warnings
-from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
-from enum import Enum
-from functools import lru_cache
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Protocol, Tuple, Union
-import unittest
-from unittest.mock import MagicMock
-
-import numpy as np
-import pandas as pd
-from numpy.typing import NDArray
-
-# Suppress warnings for cleaner output
-warnings.filterwarnings('ignore', category=pd.errors.PerformanceWarning)
-warnings.filterwarnings('ignore', category=FutureWarning)
-
-# =================== Enhanced Constants and Types ===================
-
-class SignalType(Enum):
-    """Trading signal types."""
-    LONG = 1
-    FLAT = 0
-    SHORT = -1
-
-class MarketRegime(Enum):
-    """Market regime classifications."""
-    BULL = "bull"
-    BEAR = "bear"
-    SIDEWAYS = "sideways"
-    HIGH_VOL = "high_vol"
-    LOW_VOL = "low_vol"
-
-class OrderType(Enum):
-    """Order execution types."""
-    MARKET = "market"
-    LIMIT = "limit"
-    STOP = "stop"
-    STOP_LIMIT = "stop_limit"
-
-@dataclass(frozen=True)
-class EngineConfig:
-    """Configuration for trading engine."""
-    max_position_size: float = 1.0
-    max_portfolio_risk: float = 0.02
-    enable_stop_loss: bool = True
-    enable_take_profit: bool = True
-    enable_position_sizing: bool = True
-    rebalance_frequency: str = 'daily'
-    risk_free_rate: float = 0.02
-    max_drawdown_threshold: float = 0.15
-    volatility_threshold: float = 0.30
-    min_data_points: int = 252
-    cache_indicators: bool = True
-    log_level: str = 'INFO'
-
-@dataclass(frozen=True)
-class TechnicalIndicators:
-    """Immutable container for technical indicators with validation."""
-    rsi: pd.Series
-    ema_fast: pd.Series
-    ema_slow: pd.Series
-    sma_trend: pd.Series
-    atr: pd.Series
-    bollinger_upper: pd.Series
-    bollinger_lower: pd.Series
-    bollinger_middle: pd.Series
-    macd: pd.Series
-    macd_signal: pd.Series
-    stochastic_k: pd.Series
-    stochastic_d: pd.Series
-    
-    def __post_init__(self):
-        """Validate indicators after creation."""
-        indicators = [self.rsi, self.ema_fast, self.ema_slow, self.sma_trend, 
-                     self.atr, self.bollinger_upper, self.bollinger_lower, 
-                     self.bollinger_middle, self.macd, self.macd_signal,
-                     self.stochastic_k, self.stochastic_d]
-        
-        if not all(isinstance(ind, pd.Series) for ind in indicators):
-            raise ValueError("All indicators must be pandas Series")
-        
-        # Check index alignment
-        base_index = self.rsi.index
-        if not all(ind.index.equals(base_index) for ind in indicators):
-            raise ValueError("All indicators must have the same index")
-
-@dataclass(frozen=True)
-class Trade:
-    """Individual trade record."""
-    entry_time: pd.Timestamp
-    exit_time: pd.Timestamp
-    entry_price: float
-    exit_price: float
-    position: int  # 1 for long, -1 for short
-    quantity: float
-    pnl: float
-    pnl_pct: float
-    duration: pd.Timedelta
-    exit_reason: str  # 'signal', 'stop_loss', 'take_profit'
-    commission: float = 0.0
-
-@dataclass(frozen=True)
-class BacktestResult:
-    """Enhanced immutable backtest results."""
-    equity_curve: pd.Series
-    trades: pd.DataFrame
-    daily_returns: pd.Series
-    total_return: float
-    annualized_return: float
-    volatility: float
-    sharpe_ratio: float
-    sortino_ratio: float
-    calmar_ratio: float
-    omega_ratio: float
-    max_drawdown: float
-    max_drawdown_duration: pd.Timedelta
-    win_rate: float
-    profit_factor: float
-    tail_ratio: float
-    var_95: float
-    expected_shortfall_95: float
-    total_trades: int
-    avg_trade_duration: pd.Timedelta
-    largest_win: float
-    largest_loss: float
-
-@dataclass(frozen=True)
-class TradingParameters:
-    """Enhanced immutable trading parameters with validation."""
-    # Technical indicators
-    rsi_period: int = 14
-    rsi_oversold: float = 30.0
-    rsi_overbought: float = 70.0
-    ema_fast_period: int = 12
-    ema_slow_period: int = 26
-    sma_trend_period: int = 50
-    bb_period: int = 20
-    bb_std_mult: float = 2.0
-    atr_period: int = 14
-    macd_signal_period: int = 9
-    stoch_k_period: int = 14
-    stoch_d_period: int = 3
-    
-    # Signal generation
-    signal_threshold: float = 0.1
-    ensemble_weights: Dict[str, float] = field(default_factory=lambda: {
-        'trend': 0.3, 'mean_reversion': 0.2, 'momentum': 0.3, 'volatility_breakout': 0.2
-    })
-    
-    # Risk management
-    stop_loss_atr_mult: float = 2.0
-    take_profit_atr_mult: float = 3.0
-    position_size: float = 1.0
-    max_position_risk: float = 0.02
-    
-    # Position sizing
-    volatility_target: float = 0.15
-    kelly_fraction: float = 0.25
-    
-    def __post_init__(self):
-        """Validate parameters after creation."""
-        if self.rsi_period < 2 or self.rsi_period > 50:
-            raise ValueError("RSI period must be between 2 and 50")
-        if self.ema_fast_period >= self.ema_slow_period:
-            raise ValueError("Fast EMA period must be less than slow EMA period")
-        if not 0 < self.signal_threshold < 1:
-            raise ValueError("Signal threshold must be between 0 and 1")
-        if self.stop_loss_atr_mult < 0 or self.take_profit_atr_mult < 0:
-            raise ValueError("ATR multipliers must be positive")
-        if not abs(sum(self.ensemble_weights.values()) - 1.0) < 0.001:
-            raise ValueError("Ensemble weights must sum to 1.0")
-
-# =================== Enhanced Custom Exceptions ===================
-
-class TradingEngineError(Exception):
-    """Base exception for trading engine."""
-    def __init__(self, message: str, error_code: Optional[str] = None):
-        super().__init__(message)
-        self.error_code = error_code
-        self.timestamp = pd.Timestamp.now()
-
-class DataValidationError(TradingEngineError):
-    """Data validation failed."""
-    pass
-
-class IndicatorComputationError(TradingEngineError):
-    """Indicator computation failed."""
-    pass
-
-class BacktestExecutionError(TradingEngineError):
-    """Backtest execution failed."""
-    pass
-
-class InsufficientDataError(TradingEngineError):
-    """Insufficient data for analysis."""
-    pass
-
-class ConfigurationError(TradingEngineError):
-    """Configuration error."""
-    pass
-
-class RiskLimitExceededError(TradingEngineError):
-    """Risk limit exceeded."""
-    pass
-
-# =================== Enhanced Protocols (Interfaces) ===================
-
-class DataValidator(Protocol):
-    """Protocol for data validation."""
-    
-    def validate_ohlcv(self, data: pd.DataFrame) -> pd.DataFrame:
-        """Validate and clean OHLCV data."""
-        ...
-
-class IndicatorCalculator(Protocol):
-    """Protocol for indicator calculations."""
-    
-    def calculate_indicators(self, data: pd.DataFrame, params: TradingParameters) -> TechnicalIndicators:
-        """Calculate technical indicators."""
-        ...
-
-class SignalGenerator(Protocol):
-    """Protocol for signal generation."""
-    
-    def generate_signals(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Generate trading signals."""
-        ...
-
-class RiskManager(Protocol):
-    """Protocol for risk management."""
-    
-    def apply_risk_management(self, data: pd.DataFrame, signals: pd.Series, 
-                            indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Apply risk management rules."""
-        ...
-
-class BacktestEngine(Protocol):
-    """Protocol for backtesting."""
-    
-    def run_backtest(self, data: pd.DataFrame, signals: pd.Series, 
-                    params: TradingParameters, config: EngineConfig) -> BacktestResult:
-        """Run backtest simulation."""
-        ...
-
-# =================== Enhanced Mathematical Functions ===================
-
-class MathUtils:
-    """Vectorized mathematical utilities for maximum performance."""
-    
-    @staticmethod
-    @lru_cache(maxsize=256)
-    def ema_alpha(span: int) -> float:
-        """Calculate EMA smoothing factor with caching."""
-        if span <= 0:
-            raise ValueError("EMA span must be positive")
-        return 2.0 / (span + 1.0)
-    
-    @staticmethod
-    def safe_divide(numerator: NDArray, denominator: NDArray, fill_value: float = 0.0) -> NDArray:
-        """Safe division with comprehensive NaN handling."""
-        with np.errstate(divide='ignore', invalid='ignore'):
-            result = np.divide(numerator, denominator)
-            mask = np.isfinite(result)
-            return np.where(mask, result, fill_value)
-    
-    @staticmethod
-    def rolling_apply_numba(series: pd.Series, window: int, func: callable) -> pd.Series:
-        """Ultra-fast rolling apply using numpy operations."""
-        if len(series) < window:
-            return pd.Series(index=series.index, dtype=float)
-        
-        values = series.values
-        result = np.full(len(values), np.nan)
-        
-        # Vectorized rolling computation
-        for i in range(window - 1, len(values)):
-            result[i] = func(values[i - window + 1:i + 1])
-        
-        return pd.Series(result, index=series.index)
-    
-    @staticmethod
-    def calculate_drawdown_vectorized(equity: pd.Series) -> Tuple[pd.Series, float, pd.Timedelta]:
-        """Calculate drawdown metrics using vectorized operations."""
-        peak = equity.expanding().max()
-        drawdown = (equity - peak) / peak
-        max_dd = drawdown.min()
-        
-        # Calculate max drawdown duration
-        is_peak = equity == peak
-        peak_indices = equity.index[is_peak]
-        
-        if len(peak_indices) > 1:
-            max_duration = pd.Timedelta(0)
-            for i in range(len(peak_indices) - 1):
-                duration = peak_indices[i + 1] - peak_indices[i]
-                max_duration = max(max_duration, duration)
-        else:
-            max_duration = pd.Timedelta(0)
-        
-        return drawdown, max_dd, max_duration
-
-# =================== Enhanced Data Validation Service ===================
-
-class DataValidationService:
-    """Enhanced service for validating and cleaning trading data."""
-    
-    def __init__(self, logger: logging.Logger, config: EngineConfig):
-        self._logger = logger
-        self._config = config
-    
-    def validate_ohlcv(self, data: pd.DataFrame) -> pd.DataFrame:
-        """
-        Validate and clean OHLCV data with comprehensive checks.
-        
-        Args:
-            data: DataFrame with OHLCV columns
-            
-        Returns:
-            Cleaned and validated DataFrame
-            
-        Raises:
-            DataValidationError: If data validation fails
-            InsufficientDataError: If insufficient data points
-        """
-        if data.empty:
-            raise DataValidationError("DataFrame cannot be empty")
-        
-        if len(data) < self._config.min_data_points:
-            raise InsufficientDataError(
-                f"Insufficient data: {len(data)} rows, minimum {self._config.min_data_points} required"
-            )
-        
-        required_columns = {'open', 'high', 'low', 'close', 'volume'}
-        missing_columns = required_columns - set(data.columns)
-        if missing_columns:
-            raise DataValidationError(f"Missing required columns: {missing_columns}")
-        
-        if not isinstance(data.index, pd.DatetimeIndex):
-            try:
-                data.index = pd.to_datetime(data.index)
-            except Exception as e:
-                raise DataValidationError(f"Cannot convert index to DatetimeIndex: {e}")
-        
-        # Clean and validate data
-        cleaned_data = data.copy()
-        
-        # Remove duplicates and sort
-        initial_length = len(cleaned_data)
-        cleaned_data = cleaned_data[~cleaned_data.index.duplicated(keep='last')].sort_index()
-        
-        if len(cleaned_data) < initial_length:
-            self._logger.warning(f"Removed {initial_length - len(cleaned_data)} duplicate timestamps")
-        
-        # Validate and correct OHLC relationships
-        cleaned_data = self._fix_ohlc_relationships(cleaned_data)
-        
-        # Handle missing values with forward/backward fill
-        null_counts = cleaned_data[list(required_columns)].isnull().sum()
-        if null_counts.any():
-            self._logger.warning(f"Found null values: {null_counts.to_dict()}")
-            cleaned_data[list(required_columns)] = cleaned_data[list(required_columns)].ffill().bfill()
-        
-        # Validate volume
-        if (cleaned_data['volume'] < 0).any():
-            self._logger.warning("Found negative volume values, taking absolute values")
-            cleaned_data['volume'] = cleaned_data['volume'].abs()
-        
-        # Detect and handle price anomalies
-        cleaned_data = self._detect_price_anomalies(cleaned_data)
-        
-        return cleaned_data
-    
-    def _fix_ohlc_relationships(self, data: pd.DataFrame) -> pd.DataFrame:
-        """Fix invalid OHLC relationships."""
-        data = data.copy()
-        
-        # Vectorized OHLC validation
-        max_price = data[['open', 'low', 'close']].max(axis=1)
-        min_price = data[['open', 'high', 'close']].min(axis=1)
-        
-        invalid_high = data['high'] < max_price
-        invalid_low = data['low'] > min_price
-        
-        if invalid_high.any() or invalid_low.any():
-            total_invalid = invalid_high.sum() + invalid_low.sum()
-            self._logger.warning(f"Correcting {total_invalid} invalid OHLC relationships")
-            
-            data.loc[invalid_high, 'high'] = max_price[invalid_high]
-            data.loc[invalid_low, 'low'] = min_price[invalid_low]
-        
-        return data
-    
-    def _detect_price_anomalies(self, data: pd.DataFrame) -> pd.DataFrame:
-        """Detect and handle price anomalies using statistical methods."""
-        data = data.copy()
-        
-        for col in ['open', 'high', 'low', 'close']:
-            returns = data[col].pct_change()
-            z_scores = np.abs((returns - returns.mean()) / returns.std())
-            
-            # Mark extreme outliers (>5 standard deviations)
-            outliers = z_scores > 5
-            if outliers.any():
-                self._logger.warning(f"Found {outliers.sum()} outliers in {col}, applying smoothing")
-                # Replace with rolling median
-                data.loc[outliers, col] = data[col].rolling(5, center=True).median()[outliers]
-        
-        return data
-
-# =================== Enhanced Technical Indicators Service ===================
-
-class TechnicalIndicatorsService:
-    """Optimized service for calculating technical indicators with caching."""
-    
-    def __init__(self, logger: logging.Logger, config: EngineConfig):
-        self._logger = logger
-        self._config = config
-        self._math = MathUtils()
-        self._cache: Dict[str, TechnicalIndicators] = {}
-    
-    def calculate_indicators(self, data: pd.DataFrame, params: TradingParameters) -> TechnicalIndicators:
-        """
-        Calculate all technical indicators efficiently with optional caching.
-        
-        Args:
-            data: OHLCV DataFrame
-            params: Trading parameters
-            
-        Returns:
-            Technical indicators container
-            
-        Raises:
-            IndicatorComputationError: If indicator calculation fails
-        """
-        # Check cache if enabled
-        if self._config.cache_indicators:
-            cache_key = self._get_cache_key(data, params)
-            if cache_key in self._cache:
-                self._logger.debug("Using cached indicators")
-                return self._cache[cache_key]
-        
-        try:
-            close = data['close']
-            high = data['high']
-            low = data['low']
-            
-            # Calculate all indicators using vectorized operations
-            indicators = TechnicalIndicators(
-                rsi=self._calculate_rsi_optimized(close, params.rsi_period),
-                ema_fast=self._calculate_ema_optimized(close, params.ema_fast_period),
-                ema_slow=self._calculate_ema_optimized(close, params.ema_slow_period),
-                sma_trend=self._calculate_sma_optimized(close, params.sma_trend_period),
-                atr=self._calculate_atr_optimized(high, low, close, params.atr_period),
-                **self._calculate_bollinger_bands_optimized(close, params.bb_period, params.bb_std_mult),
-                **self._calculate_macd_optimized(close, params.ema_fast_period, 
-                                               params.ema_slow_period, params.macd_signal_period),
-                **self._calculate_stochastic_optimized(high, low, close, 
-                                                     params.stoch_k_period, params.stoch_d_period)
-            )
-            
-            # Cache if enabled
-            if self._config.cache_indicators:
-                self._cache[cache_key] = indicators
-                # Limit cache size
-                if len(self._cache) > 50:
-                    oldest_key = next(iter(self._cache))
-                    del self._cache[oldest_key]
-            
-            return indicators
-            
-        except Exception as e:
-            raise IndicatorComputationError(f"Failed to calculate indicators: {e}") from e
-    
-    def _get_cache_key(self, data: pd.DataFrame, params: TradingParameters) -> str:
-        """Generate cache key for indicators."""
-        # Use hash of data characteristics and parameters
-        data_hash = hashlib.md5(
-            (str(data.index[0]) + str(data.index[-1]) + str(len(data))).encode()
-        ).hexdigest()[:8]
-        
-        params_hash = hashlib.md5(str(params).encode()).hexdigest()[:8]
-        return f"{data_hash}_{params_hash}"
-    
-    def _calculate_rsi_optimized(self, series: pd.Series, period: int) -> pd.Series:
-        """Calculate RSI using optimized vectorized operations."""
-        if len(series) < period + 1:
-            return pd.Series(50.0, index=series.index)
-        
-        delta = series.diff()
-        gain = delta.where(delta > 0, 0.0)
-        loss = -delta.where(delta < 0, 0.0)
-        
-        # Use Wilder's smoothing (equivalent to EMA with specific alpha)
-        alpha = 1.0 / period
-        avg_gain = gain.ewm(alpha=alpha, adjust=False).mean()
-        avg_loss = loss.ewm(alpha=alpha, adjust=False).mean()
-        
-        rs = self._math.safe_divide(avg_gain.values, avg_loss.values, 100.0)
-        rsi = 100.0 - (100.0 / (1.0 + rs))
-        
-        return pd.Series(rsi, index=series.index).fillna(50.0)
-    
-    def _calculate_ema_optimized(self, series: pd.Series, span: int) -> pd.Series:
-        """Calculate EMA with input validation."""
-        if span <= 0:
-            raise ValueError("EMA span must be positive")
-        if len(series) < span:
-            return pd.Series(index=series.index, dtype=float)
-        
-        return series.ewm(span=span, adjust=False).mean()
-    
-    def _calculate_sma_optimized(self, series: pd.Series, window: int) -> pd.Series:
-        """Calculate SMA with validation."""
-        if window <= 0:
-            raise ValueError("SMA window must be positive")
-        if len(series) < window:
-            return pd.Series(index=series.index, dtype=float)
-        
-        return series.rolling(window=window, min_periods=1).mean()
-    
-    def _calculate_atr_optimized(self, high: pd.Series, low: pd.Series, 
-                               close: pd.Series, period: int) -> pd.Series:
-        """Calculate Average True Range with enhanced accuracy."""
-        if period <= 0:
-            raise ValueError("ATR period must be positive")
-        
-        # True Range calculation
-        high_low = high - low
-        high_close_prev = (high - close.shift(1)).abs()
-        low_close_prev = (low - close.shift(1)).abs()
-        
-        # Use pandas concat for better performance
-        ranges = pd.concat([high_low, high_close_prev, low_close_prev], axis=1)
-        true_range = ranges.max(axis=1)
-        
-        # Use Wilder's smoothing
-        alpha = 1.0 / period
-        return true_range.ewm(alpha=alpha, adjust=False).mean()
-    
-    def _calculate_bollinger_bands_optimized(self, series: pd.Series, 
-                                           period: int, std_mult: float) -> Dict[str, pd.Series]:
-        """Calculate Bollinger Bands with validation."""
-        if period <= 0:
-            raise ValueError("Bollinger Bands period must be positive")
-        
-        sma = self._calculate_sma_optimized(series, period)
-        std = series.rolling(window=period, min_periods=1).std()
-        
-        return {
-            'bollinger_upper': sma + (std * std_mult),
-            'bollinger_lower': sma - (std * std_mult),
-            'bollinger_middle': sma
-        }
-    
-    def _calculate_macd_optimized(self, series: pd.Series, fast_period: int, 
-                                slow_period: int, signal_period: int) -> Dict[str, pd.Series]:
-        """Calculate MACD with validation."""
-        if fast_period >= slow_period:
-            raise ValueError("Fast period must be less than slow period")
-        
-        ema_fast = self._calculate_ema_optimized(series, fast_period)
-        ema_slow = self._calculate_ema_optimized(series, slow_period)
-        macd = ema_fast - ema_slow
-        signal = self._calculate_ema_optimized(macd, signal_period)
-        
-        return {
-            'macd': macd,
-            'macd_signal': signal
-        }
-    
-    def _calculate_stochastic_optimized(self, high: pd.Series, low: pd.Series, 
-                                      close: pd.Series, k_period: int, d_period: int) -> Dict[str, pd.Series]:
-        """Calculate Stochastic Oscillator with validation."""
-        if k_period <= 0 or d_period <= 0:
-            raise ValueError("Stochastic periods must be positive")
-        
-        lowest_low = low.rolling(window=k_period, min_periods=1).min()
-        highest_high = high.rolling(window=k_period, min_periods=1).max()
-        
-        # Avoid division by zero
-        denominator = highest_high - lowest_low
-        k_percent = ((close - lowest_low) / denominator).fillna(0) * 100
-        k_percent = k_percent.clip(0, 100)  # Ensure bounds
-        
-        d_percent = k_percent.rolling(window=d_period, min_periods=1).mean()
-        
-        return {
-            'stochastic_k': k_percent,
-            'stochastic_d': d_percent
-        }
-
-# =================== Enhanced Signal Generation Service ===================
-
-class TradingSignalService:
-    """Enhanced service for generating trading signals using ensemble methods."""
-    
-    def __init__(self, logger: logging.Logger):
-        self._logger = logger
-    
-    def generate_signals(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """
-        Generate ensemble trading signals with improved logic.
-        
-        Args:
-            indicators: Technical indicators
-            params: Trading parameters
-            
-        Returns:
-            Series with trading signals (-1, 0, 1)
-        """
-        # Generate individual signals
-        signals_dict = {
-            'trend': self._trend_following_signal(indicators, params),
-            'mean_reversion': self._mean_reversion_signal(indicators, params),
-            'momentum': self._momentum_signal(indicators, params),
-            'volatility_breakout': self._volatility_breakout_signal(indicators, params)
-        }
-        
-        # Combine using ensemble weights
-        ensemble_signal = pd.Series(0.0, index=indicators.rsi.index)
-        for signal_name, signal in signals_dict.items():
-            weight = params.ensemble_weights.get(signal_name, 0.0)
-            ensemble_signal += signal * weight
-        
-        # Apply threshold and generate final signals
-        final_signals = pd.Series(SignalType.FLAT.value, index=ensemble_signal.index, dtype=int)
-        final_signals[ensemble_signal > params.signal_threshold] = SignalType.LONG.value
-        final_signals[ensemble_signal < -params.signal_threshold] = SignalType.SHORT.value
-        
-        # Apply signal smoothing to reduce noise
-        final_signals = self._smooth_signals(final_signals)
-        
-        return final_signals
-    
-    def _trend_following_signal(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Generate trend following signals with proper Series handling."""
-        # EMA crossover signal
-        ema_diff = indicators.ema_fast - indicators.ema_slow
-        ema_signal = pd.Series(
-            np.where(ema_diff > 0, 1.0, -1.0),
-            index=indicators.ema_fast.index
-        )
-        
-        # Price vs SMA trend
-        price_vs_sma = indicators.ema_fast - indicators.sma_trend
-        trend_signal = pd.Series(
-            np.where(price_vs_sma > 0, 1.0, -1.0),
-            index=indicators.ema_fast.index
-        )
-        
-        # Combine with proper weighting
-        combined_signal = ema_signal * 0.6 + trend_signal * 0.4
-        return combined_signal
-    
-    def _mean_reversion_signal(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Generate mean reversion signals with RSI and Bollinger Bands."""
-        # RSI-based signals
-        rsi_signal = pd.Series(0.0, index=indicators.rsi.index)
-        rsi_signal[indicators.rsi < params.rsi_oversold] = 1.0
-        rsi_signal[indicators.rsi > params.rsi_overbought] = -1.0
-        
-        # Bollinger Bands signals
-        bb_signal = pd.Series(0.0, index=indicators.ema_fast.index)
-        bb_signal[indicators.ema_fast < indicators.bollinger_lower] = 1.0
-        bb_signal[indicators.ema_fast > indicators.bollinger_upper] = -1.0
-        
-        # Combine signals
-        combined_signal = rsi_signal * 0.5 + bb_signal * 0.5
-        return combined_signal
-    
-    def _momentum_signal(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Generate momentum signals using MACD and Stochastic."""
-        # MACD signal
-        macd_signal = pd.Series(
-            np.where(indicators.macd > indicators.macd_signal, 1.0, -1.0),
-            index=indicators.macd.index
-        )
-        
-        # Stochastic signal with overbought/oversold levels
-        stoch_signal = pd.Series(0.0, index=indicators.stochastic_k.index)
-        
-        # Buy signal when %K crosses above %D and not overbought
-        buy_condition = (
-            (indicators.stochastic_k > indicators.stochastic_d) & 
-            (indicators.stochastic_k < 80)
-        )
-        
-        # Sell signal when %K crosses below %D and not oversold
-        sell_condition = (
-            (indicators.stochastic_k < indicators.stochastic_d) & 
-            (indicators.stochastic_k > 20)
-        )
-        
-        stoch_signal[buy_condition] = 1.0
-        stoch_signal[sell_condition] = -1.0
-        
-        # Combine signals
-        combined_signal = macd_signal * 0.6 + stoch_signal * 0.4
-        return combined_signal
-    
-    def _volatility_breakout_signal(self, indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """Generate volatility breakout signals."""
-        # ATR-based volatility measure
-        atr_pct = indicators.atr / indicators.ema_fast
-        atr_threshold = atr_pct.rolling(20).quantile(0.8)
-        high_vol = atr_pct > atr_threshold
-        
-        # Bollinger Band width expansion
-        bb_width = (indicators.bollinger_upper - indicators.bollinger_lower) / indicators.bollinger_middle
-        bb_expanding = bb_width > bb_width.rolling(20).mean()
-        
-        # Trend direction
-        trend_up = indicators.ema_fast > indicators.ema_slow
-        trend_down = indicators.ema_fast < indicators.ema_slow
-        
-        # Generate breakout signals
-        breakout_signal = pd.Series(0.0, index=indicators.atr.index)
-        
-        long_condition = high_vol & bb_expanding & trend_up
-        short_condition = high_vol & bb_expanding & trend_down
-        
-        breakout_signal[long_condition] = 1.0
-        breakout_signal[short_condition] = -1.0
-        
-        return breakout_signal
-    
-    def _smooth_signals(self, signals: pd.Series, window: int = 3) -> pd.Series:
-        """Smooth signals to reduce noise and whipsaws."""
-        # Use modal smoothing to reduce signal noise
-        smoothed = signals.rolling(window=window, center=True).apply(
-            lambda x: x.mode().iloc[0] if len(x.mode()) > 0 else x.iloc[len(x)//2], 
-            raw=False
-        ).fillna(signals)
-        
-        return smoothed.astype(int)
-
-# =================== Enhanced Risk Management Service ===================
-
-class RiskManagementService:
-    """Comprehensive risk management with stop-loss, take-profit, and position sizing."""
-    
-    def __init__(self, logger: logging.Logger):
-        self._logger = logger
-    
-    def apply_risk_management(self, data: pd.DataFrame, signals: pd.Series, 
-                            indicators: TechnicalIndicators, params: TradingParameters) -> pd.Series:
-        """
-        Apply comprehensive risk management including stop-loss and take-profit.
-        
-        Args:
-            data: OHLCV DataFrame
-            signals: Raw trading signals
-            indicators: Technical indicators
-            params: Trading parameters
-            
-        Returns:
-            Risk-managed signals
-        """
-        managed_signals = signals.copy()
-        current_position = 0
-        entry_price = 0.0
-        entry_time = None
-        
-        for i in range(len(signals)):
-            timestamp = signals.index[i]
-            current_signal = signals.iloc[i]
-            current_price = data.loc[timestamp, 'close']
-            current_atr = indicators.atr.iloc[i]
-            
-            # Check for position exit conditions first
-            if current_position != 0 and entry_price > 0:
-                exit_signal = self._check_exit_conditions(
-                    current_position, entry_price, current_price, current_atr, params
-                )
-                
-                if exit_signal:
-                    managed_signals.iloc[i] = 0  # Force exit
-                    self._logger.debug(f"Risk management exit at {timestamp}: {exit_signal}")
-                    current_position = 0
-                    entry_price = 0.0
-                    entry_time = None
-                    continue
-            
-            # Handle new position entries
-            if current_signal != 0 and current_position == 0:
-                # Apply position sizing
-                position_size = self._calculate_position_size(
-                    current_price, current_atr, params
-                )
-                
-                if position_size > 0:
-                    current_position = current_signal
-                    entry_price = current_price
-                    entry_time = timestamp
-                    managed_signals.iloc[i] = current_signal
-                else:
-                    managed_signals.iloc[i] = 0  # Risk too high, skip entry
-                    
-            elif current_signal == 0 and current_position != 0:
-                # Natural signal exit
-                current_position = 0
-                entry_price = 0.0
-                entry_time = None
-                managed_signals.iloc[i] = 0
-                
-            elif current_position != 0:
-                # Maintain current position
-                managed_signals.iloc[i] = current_position
-            else:
-                # No position, no signal
-                managed_signals.iloc[i] = 0
-        
-        return managed_signals
-    
-    def _check_exit_conditions(self, position: int, entry_price: float, 
-                             current_price: float, atr: float, params: TradingParameters) -> Optional[str]:
-        """Check if position should be exited based on risk management rules."""
-        if pd.isna(atr) or atr == 0:
-            return None
-        
-        if position > 0:  # Long position
-            # Stop loss
-            stop_loss_price = entry_price - (atr * params.stop_loss_atr_mult)
-            if current_price <= stop_loss_price:
-                return "stop_loss"
-            
-            # Take profit
-            take_profit_price = entry_price + (atr * params.take_profit_atr_mult)
-            if current_price >= take_profit_price:
-                return "take_profit"
-                
-        elif position < 0:  # Short position
-            # Stop loss
-            stop_loss_price = entry_price + (atr * params.stop_loss_atr_mult)
-            if current_price >= stop_loss_price:
-                return "stop_loss"
-            
-            # Take profit
-            take_profit_price = entry_price - (atr * params.take_profit_atr_mult)
-            if current_price <= take_profit_price:
-                return "take_profit"
-        
-        return None
-    
-    def _calculate_position_size(self, price: float, atr: float, params: TradingParameters) -> float:
-        """Calculate position size based on volatility and risk parameters."""
-        if pd.isna(atr) or atr == 0 or price == 0:
-            return 0.0
-        
-        # Risk per share based on ATR stop loss
-        risk_per_share = atr * params.stop_loss_atr_mult
-        risk_percentage = risk_per_share / price
-        
-        # Limit position size based on maximum risk
-        if risk_percentage > params.max_position_risk:
-            return 0.0
-        
-        # Volatility-based position sizing
-        volatility_adjusted_size = params.volatility_target / risk_percentage
-        
-        # Apply Kelly fraction for position sizing
-        final_size = min(
-            volatility_adjusted_size * params.kelly_fraction,
-            params.position_size
-        )
-        
-        return max(0.0, final_size)
-
-# =================== Enhanced Vectorized Backtesting Engine ===================
-
-class VectorizedBacktestEngine:
-    """Ultra-high-performance vectorized backtesting engine with comprehensive metrics."""
-    
-    def __init__(self, logger: logging.Logger):
-        self._logger = logger
-        self._math = MathUtils()
-    
-    def run_backtest(self, data: pd.DataFrame, signals: pd.Series, 
-                    params: TradingParameters, config: EngineConfig, 
-                    initial_capital: float = 10000.0, fee_bps: float = 5.0) -> BacktestResult:
-        """
-        Run comprehensive vectorized backtest simulation.
-        
-        Args:
-            data: OHLCV DataFrame
-            signals: Trading signals
-            params: Trading parameters
-            config: Engine configuration
-            initial_capital: Starting capital
-            fee_bps: Transaction fees in basis points
-            
-        Returns:
-            Comprehensive backtest results
-            
-        Raises:
-            BacktestExecutionError: If backtest execution fails
-        """
-        try:
-            # Align data and signals
-            aligned_data = data.reindex(signals.index, method='ffill')
-            
-            # Calculate returns and equity curve
-            returns = self._calculate_returns_vectorized(aligned_data, signals, params, fee_bps)
-            equity_curve = (1 + returns).cumprod() * initial_capital
-            
-            # Generate comprehensive trades DataFrame
-            trades_df = self._generate_trades_dataframe_vectorized(aligned_data, signals, params)
-            
-            # Calculate all performance metrics
-            metrics = self._calculate_comprehensive_metrics(equity_curve, trades_df, returns, config.risk_free_rate)
-            
-            return BacktestResult(
-                equity_curve=equity_curve,
-                trades=trades_df,
-                daily_returns=returns,
-                **metrics
-            )
-            
-        except Exception as e:
-            raise BacktestExecutionError(f"Backtest execution failed: {e}") from e
-    
-    def _calculate_returns_vectorized(self, data: pd.DataFrame, signals: pd.Series, 
-                                    params: TradingParameters, fee_bps: float) -> pd.Series:
-        """Calculate returns using fully vectorized operations."""
-        # Shift signals by 1 to avoid look-ahead bias
-        position = signals.shift(1).fillna(0)
-        
-        # Apply position sizing
-        position = position * params.position_size
-        
-        # Calculate price returns
-        price_returns = data['close'].pct_change().fillna(0)
-        
-        # Calculate strategy returns
-        strategy_returns = position * price_returns
-        
-        # Apply transaction costs vectorized
-        position_changes = position.diff().abs()
-        transaction_costs = position_changes * (fee_bps / 10000.0)
-        
-        # Net returns
-        net_returns = strategy_returns - transaction_costs
-        
-        return net_returns
-    
-    def _generate_trades_dataframe_vectorized(self, data: pd.DataFrame, signals: pd.Series, 
-                                            params: TradingParameters) -> pd.DataFrame:
-        """Generate trades DataFrame using optimized vectorized operations."""
-        # Find signal changes
-        signal_changes = signals.diff().fillna(signals)
-        trade_points = signal_changes != 0
-        
-        if not trade_points.any():
-            return pd.DataFrame()
-        
-        trades = []
-        position = 0
-        entry_idx = None
-        
-        for i, (timestamp, signal) in enumerate(signals.items()):
-            if timestamp not in data.index:
-                continue
-                
-            current_price = data.loc[timestamp, 'close']
-            
-            # Position change logic
-            if signal != position:
-                # Close existing position
-                if position != 0 and entry_idx is not None:
-                    entry_timestamp = signals.index[entry_idx]
-                    entry_price = data.loc[entry_timestamp, 'close']
-                    
-                    # Calculate trade metrics
-                    pnl = (current_price - entry_price) * position
-                    pnl_pct = pnl / entry_price
-                    duration = timestamp - entry_timestamp
-                    
-                    # Determine exit reason
-                    if signal == 0:
-                        exit_reason = "signal"
-                    else:
-                        exit_reason = "reversal"
-                    
-                    trades.append({
-                        'entry_time': entry_timestamp,
-                        'exit_time': timestamp,
-                        'entry_price': entry_price,
-                        'exit_price': current_price,
-                        'position': position,
-                        'quantity': abs(position) * params.position_size,
-                        'pnl': pnl,
-                        'pnl_pct': pnl_pct,
-                        'duration': duration,
-                        'exit_reason': exit_reason,
-                        'commission': abs(position) * params.position_size * current_price * 0.0005  # 5 bps
-                    })
-                
-                # Open new position
-                if signal != 0:
-                    position = signal
-                    entry_idx = i
-                else:
-                    position = 0
-                    entry_idx = None
-        
-        return pd.DataFrame(trades)
-    
-    def _calculate_comprehensive_metrics(self, equity_curve: pd.Series, trades_df: pd.DataFrame, 
-                                       returns: pd.Series, risk_free_rate: float) -> Dict[str, Any]:
-        """Calculate comprehensive performance metrics."""
-        if equity_curve.empty or len(equity_curve) < 2:
-            return self._empty_comprehensive_metrics()
-        
-        initial_capital = equity_curve.iloc[0]
-        final_capital = equity_curve.iloc[-1]
-        
-        # Basic return metrics
-        total_return = (final_capital / initial_capital) - 1.0
-        n_years = len(equity_curve) / 252  # Assuming daily data
-        annualized_return = (1 + total_return) ** (1/n_years) - 1.0 if n_years > 0 else 0.0
-        
-        # Risk metrics
-        volatility = returns.std() * np.sqrt(252)
-        
-        # Drawdown analysis
-        drawdown, max_drawdown, max_dd_duration = self._math.calculate_drawdown_vectorized(equity_curve)
-        
-        # Risk-adjusted returns
-        excess_returns = returns - (risk_free_rate / 252)  # Daily risk-free rate
-        sharpe_ratio = (excess_returns.mean() / returns.std()) * np.sqrt(252) if returns.std() > 0 else 0.0
-        
-        # Sortino ratio
-        downside_returns = returns[returns < 0]
-        downside_std = downside_returns.std() * np.sqrt(252)
-        sortino_ratio = (returns.mean() * 252) / downside_std if downside_std > 0 else 0.0
-        
-        # Calmar ratio
-        calmar_ratio = annualized_return / abs(max_drawdown) if max_drawdown < 0 else 0.0
-        
-        # Omega ratio
-        threshold = 0.0
-        positive_returns = returns[returns > threshold]
-        negative_returns = returns[returns <= threshold]
-        omega_ratio = positive_returns.sum() / abs(negative_returns.sum()) if len(negative_returns) > 0 else float('inf')
-        
-        # VaR and Expected Shortfall
-        var_95 = returns.quantile(0.05)
-        expected_shortfall_95 = returns[returns <= var_95].mean() if (returns <= var_95).any() else 0.0
-        
-        # Tail ratio
-        tail_ratio = returns.quantile(0.95) / abs(returns.quantile(0.05)) if returns.quantile(0.05) != 0 else 0.0
-        
-        # Trade-based metrics
-        if not trades_df.empty:
-            winning_trades = trades_df[trades_df['pnl'] > 0]
-            losing_trades = trades_df[trades_df['pnl'] < 0]
-            
-            win_rate = len(winning_trades) / len(trades_df)
-            
-            gross_profit = winning_trades['pnl'].sum() if not winning_trades.empty else 0.0
-            gross_loss = abs(losing_trades['pnl'].sum()) if not losing_trades.empty else 0.0
-            profit_factor = gross_profit / gross_loss if gross_loss > 0 else float('inf')
-            
-            avg_trade_duration = trades_df['duration'].mean()
-            largest_win = trades_df['pnl'].max()
-            largest_loss = trades_df['pnl'].min()
-            total_trades = len(trades_df)
-        else:
-            win_rate = 0.0
-            profit_factor = 0.0
-            avg_trade_duration = pd.Timedelta(0)
-            largest_win = 0.0
-            largest_loss = 0.0
-            total_trades = 0
-        
-        return {
-            'total_return': total_return,
-            'annualized_return': annualized_return,
-            'volatility': volatility,
-            'sharpe_ratio': sharpe_ratio,
-            'sortino_ratio': sortino_ratio,
-            'calmar_ratio': calmar_ratio,
-            'omega_ratio': omega_ratio,
-            'max_drawdown': max_drawdown,
-            'max_drawdown_duration': max_dd_duration,
-            'win_rate': win_rate,
-            'profit_factor': profit_factor,
-            'tail_ratio': tail_ratio,
-            'var_95': var_95,
-            'expected_shortfall_95': expected_shortfall_95,
-            'total_trades': total_trades,
-            'avg_trade_duration': avg_trade_duration,
-            'largest_win': largest_win,
-            'largest_loss': largest_loss
-        }
-    
-    def _empty_comprehensive_metrics(self) -> Dict[str, Any]:
-        """Return empty comprehensive metrics dictionary."""
-        return {
-            'total_return': 0.0,
-            'annualized_return': 0.0,
-            'volatility': 0.0,
-            'sharpe_ratio': 0.0,
-            'sortino_ratio': 0.0,
-            'calmar_ratio': 0.0,
-            'omega_ratio': 0.0,
-            'max_drawdown': 0.0,
-            'max_drawdown_duration': pd.Timedelta(0),
-            'win_rate': 0.0,
-            'profit_factor': 0.0,
-            'tail_ratio': 0.0,
-            'var_95': 0.0,
-            'expected_shortfall_95': 0.0,
-            'total_trades': 0,
-            'avg_trade_duration': pd.Timedelta(0),
-            'largest_win': 0.0,
-            'largest_loss': 0.0
-        }
-
-# =================== Enhanced Main Trading Engine ===================
-
-class TradingEngine:
-    """Enhanced main trading engine with comprehensive dependency injection."""
-    
-    def __init__(self, 
-                 config: Optional[EngineConfig] = None,
-                 validator: Optional[DataValidator] = None,
-                 indicator_calculator: Optional[IndicatorCalculator] = None,
-                 signal_generator: Optional[SignalGenerator] = None,
-                 risk_manager: Optional[RiskManager] = None,
-                 backtest_engine: Optional[BacktestEngine] = None,
-                 logger: Optional[logging.Logger] = None):
-        """Initialize with comprehensive dependency injection."""
-        
-        self._config = config or EngineConfig()
-        self._logger = logger or self._setup_logger()
-        
-        # Inject dependencies or use enhanced defaults
-        self._validator = validator or DataValidationService(self._logger, self._config)
-        self._indicator_calculator = indicator_calculator or TechnicalIndicatorsService(self._logger, self._config)
-        self._signal_generator = signal_generator or TradingSignalService(self._logger)
-        self._risk_manager = risk_manager or RiskManagementService(self._logger)
-        self._backtest_engine = backtest_engine or VectorizedBacktestEngine(self._logger)
-        
-        self._performance_monitor = PerformanceMonitor(self._logger)
-    
-    def run_strategy(self, data: pd.DataFrame, params: TradingParameters, 
-                    initial_capital: float = 10000.0, fee_bps: float = 5.0) -> BacktestResult:
-        """
-        Run complete enhanced trading strategy with comprehensive monitoring.
-        
-        Args:
-            data: OHLCV DataFrame
-            params: Trading parameters
-            initial_capital: Starting capital
-            fee_bps: Transaction fees in basis points
-            
-        Returns:
-            Comprehensive backtest results
-            
-        Raises:
-            TradingEngineError: If strategy execution fails
-        """
-        try:
-            self._logger.info("Starting enhanced trading strategy execution")
-            start_time = pd.Timestamp.now()
-            
-            # Validate data
-            validated_data = self._validator.validate_ohlcv(data)
-            self._logger.info(f"Validated {len(validated_data)} data points")
-            
-            # Calculate indicators
-            indicators = self._indicator_calculator.calculate_indicators(validated_data, params)
-            self._logger.info("Calculated technical indicators")
-            
-            # Generate raw signals
-            raw_signals = self._signal_generator.generate_signals(indicators, params)
-            self._logger.info(f"Generated {(raw_signals != 0).sum()} raw trading signals")
-            
-            # Apply risk management
-            managed_signals = self._risk_manager.apply_risk_management(
-                validated_data, raw_signals, indicators, params
-            )
-            risk_filtered = (managed_signals != raw_signals).sum()
-            self._logger.info(f"Risk management filtered {risk_filtered} signals")
-            
-            # Run backtest
-            result = self._backtest_engine.run_backtest(
-                validated_data, managed_signals, params, self._config, initial_capital, fee_bps
-            )
-            
-            # Monitor performance
-            self._performance_monitor.monitor_drawdown(result.equity_curve, self._config.max_drawdown_threshold)
-            self._performance_monitor.monitor_volatility(result.equity_curve, self._config.volatility_threshold)
-            
-            execution_time = pd.Timestamp.now() - start_time
-            self._logger.info(
-                f"Strategy completed in {execution_time.total_seconds():.2f}s: "
-                f"{result.total_return:.2%} total return, "
-                f"{result.sharpe_ratio:.2f} Sharpe ratio"
-            )
-            
-            return result
-            
-        except Exception as e:
-            self._logger.error(f"Strategy execution failed: {e}")
-            raise TradingEngineError(f"Strategy execution failed: {e}") from e
-    
-    def optimize_parameters(self, data: pd.DataFrame, param_ranges: Dict[str, List], 
-                           objective: str = 'sharpe_ratio', max_iterations: int = 1000) -> Tuple[TradingParameters, float]:
-        """
-        Enhanced parameter optimization with smart search.
-        
-        Args:
-            data: OHLCV DataFrame
-            param_ranges: Dictionary of parameter ranges
-            objective: Optimization objective
-            max_iterations: Maximum optimization iterations
-            
-        Returns:
-            Tuple of best parameters and best score
-        """
-        best_params = None
-        best_score = float('-inf')
-        iterations = 0
-        
-        self._logger.info(f"Starting parameter optimization with objective: {objective}")
-        
-        # Smart grid search with early stopping
-        for rsi_period in param_ranges.get('rsi_period', [14]):
-            for ema_fast in param_ranges.get('ema_fast_period', [12]):
-                for ema_slow in param_ranges.get('ema_slow_period', [26]):
-                    for signal_threshold in param_ranges.get('signal_threshold', [0.1]):
-                        
-                        if iterations >= max_iterations:
-                            break
-                        
-                        if ema_fast >= ema_slow:  # Skip invalid combinations
-                            continue
-                        
-                        try:
-                            params = TradingParameters(
-                                rsi_period=rsi_period,
-                                ema_fast_period=ema_fast,
-                                ema_slow_period=ema_slow,
-                                signal_threshold=signal_threshold
-                            )
-                            
-                            result = self.run_strategy(data, params)
-                            score = getattr(result, objective)
-                            
-                            if score > best_score:
-                                best_score = score
-                                best_params = params
-                                self._logger.info(f"New best {objective}: {score:.4f}")
-                            
-                            iterations += 1
-                            
-                        except Exception as e:
-                            self._logger.warning(f"Optimization failed for params {params}: {e}")
-        
-        self._logger.info(f"Optimization completed after {iterations} iterations")
-        return best_params, best_score
-    
-    def get_performance_alerts(self) -> List[str]:
-        """Get current performance alerts."""
-        return self._performance_monitor.get_alerts()
-    
-    def _setup_logger(self) -> logging.Logger:
-        """Setup enhanced logger with configuration."""
-        logger = logging.getLogger(__name__)
-        if not logger.handlers:
-            handler = logging.StreamHandler()
-            formatter = logging.Formatter(
-                '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-            )
-            handler.setFormatter(formatter)
-            logger.addHandler(handler)
-            logger.setLevel(getattr(logging, self._config.log_level.upper()))
-        return logger
-
-# =================== Performance Monitoring ===================
-
-class PerformanceMonitor:
-    """Enhanced real-time performance monitoring and alerting."""
-    
-    def __init__(self, logger: logging.Logger):
-        self._logger = logger
-        self._alerts: List[str] = []
-    
-    def monitor_drawdown(self, equity_curve: pd.Series, max_dd_threshold: float = 0.15):
-        """Monitor for excessive drawdowns with detailed analysis."""
-        if len(equity_curve) < 2:
-            return
-        
-        peak = equity_curve.expanding().max()
-        current_dd = (equity_curve.iloc[-1] - peak.iloc[-1]) / peak.iloc[-1]
-        
-        if current_dd < -max_dd_threshold:
-            alert = (
-                f"WARNING: Current drawdown {current_dd:.2%} exceeds threshold {max_dd_threshold:.2%}. "
-                f"Peak value: ${peak.iloc[-1]:,.2f}, Current value: ${equity_curve.iloc[-1]:,.2f}"
-            )
-            self._alerts.append(alert)
-            self._logger.warning(alert)
-    
-    def monitor_volatility(self, equity_curve: pd.Series, vol_threshold: float = 0.30):
-        """Monitor for excessive volatility with rolling analysis."""
-        if len(equity_curve) < 30:
-            return
-        
-        returns = equity_curve.pct_change().dropna()
-        recent_returns = returns.tail(30)
-        current_vol = recent_returns.std() * np.sqrt(252)
-        
-        if current_vol > vol_threshold:
-            avg_vol = returns.std() * np.sqrt(252)
-            alert = (
-                f"WARNING: Recent volatility {current_vol:.2%} exceeds threshold {vol_threshold:.2%}. "
-                f"Average volatility: {avg_vol:.2%}"
-            )
-            self._alerts.append(alert)
-            self._logger.warning(alert)
-    
-    def monitor_consecutive_losses(self, trades_df: pd.DataFrame, max_consecutive: int = 5):
-        """Monitor for consecutive losing trades."""
-        if trades_df.empty:
-            return
-        
-        trades_df = trades_df.sort_values('exit_time')
-        consecutive_losses = 0
-        max_consecutive_losses = 0
-        
-        for _, trade in trades_df.iterrows():
-            if trade['pnl'] < 0:
-                consecutive_losses += 1
-                max_consecutive_losses = max(max_consecutive_losses, consecutive_losses)
-            else:
-                consecutive_losses = 0
-        
-        if max_consecutive_losses >= max_consecutive:
-            alert = f"WARNING: {max_consecutive_losses} consecutive losing trades detected"
-            self._alerts.append(alert)
-            self._logger.warning(alert)
-    
-    def get_alerts(self) -> List[str]:
-        """Get all alerts."""
-        return self._alerts.copy()
-    
-    def clear_alerts(self):
-        """Clear all alerts."""
-        self._alerts.clear()
-
-# =================== Comprehensive Testing Framework ===================
-
-class TestTradingEngine(unittest.TestCase):
-    """Comprehensive test suite for trading engine."""
-    
-    def setUp(self):
-        """Set up test fixtures."""
-        self.config = EngineConfig(log_level='ERROR')  # Suppress logs during testing
-        self.engine = TradingEngine(config=self.config)
-        self.sample_data = self._create_sample_data()
-        self.params = TradingParameters()
-    
-    def _create_sample_data(self) -> pd.DataFrame:
-        """Create sample OHLCV data for testing."""
-        dates = pd.date_range(start='2020-01-01', end='2022-01-01', freq='D')
-        np.random.seed(42)
-        
-        n_points = len(dates)
-        base_price = 100.0
-        returns = np.random.normal(0.0005, 0.02, n_points)
-        
-        prices = base_price * (1 + returns).cumprod()
-        noise = np.random.normal(0, 0.5, n_points)
-        
-        data = pd.DataFrame({
-            'open': prices + noise,
-            'close': prices,
-            'high': prices + abs(noise) + np.random.exponential(0.5, n_points),
-            'low': prices - abs(noise) - np.random.exponential(0.5, n_points),
-            'volume': np.random.randint(1000, 10000, n_points)
-        }, index=dates)
-        
-        # Ensure OHLC relationships
-        data['high'] = np.maximum.reduce([data['open'], data['high'], data['low'], data['close']])
-        data['low'] = np.minimum.reduce([data['open'], data['high'], data['low'], data['close']])
-        
-        return data
-    
-    def test_data_validation(self):
-        """Test data validation functionality."""
-        validator = DataValidationService(MagicMock(), self.config)
-        validated = validator.validate_ohlcv(self.sample_data)
-        
-        self.assertFalse(validated.empty)
-        self.assertTrue(all(col in validated.columns for col in ['open', 'high', 'low', 'close', 'volume']))
-        self.assertIsInstance(validated.index, pd.DatetimeIndex)
-    
-    def test_indicator_calculation(self):
-        """Test technical indicator calculations."""
-        calculator = TechnicalIndicatorsService(MagicMock(), self.config)
-        indicators = calculator.calculate_indicators(self.sample_data, self.params)
-        
-        self.assertIsInstance(indicators, TechnicalIndicators)
-        self.assertEqual(len(indicators.rsi), len(self.sample_data))
-        self.assertTrue(all(0 <= rsi <= 100 for rsi in indicators.rsi.dropna()))
-    
-    def test_signal_generation(self):
-        """Test signal generation."""
-        calculator = TechnicalIndicatorsService(MagicMock(), self.config)
-        generator = TradingSignalService(MagicMock())
-        
-        indicators = calculator.calculate_indicators(self.sample_data, self.params)
-        signals = generator.generate_signals(indicators, self.params)
-        
-        self.assertEqual(len(signals), len(self.sample_data))
-        self.assertTrue(all(signal in [-1, 0, 1] for signal in signals))
-    
-    def test_risk_management(self):
-        """Test risk management functionality."""
-        calculator = TechnicalIndicatorsService(MagicMock(), self.config)
-        generator = TradingSignalService(MagicMock())
-        risk_manager = RiskManagementService(MagicMock())
-        
-        indicators = calculator.calculate_indicators(self.sample_data, self.params)
-        raw_signals = generator.generate_signals(indicators, self.params)
-        managed_signals = risk_manager.apply_risk_management(
-            self.sample_data, raw_signals, indicators, self.params
-        )
-        
-        self.assertEqual(len(managed_signals), len(raw_signals))
-        self.assertTrue(all(signal in [-1, 0, 1] for signal in managed_signals))
-    
-    def test_backtest_execution(self):
-        """Test backtest execution."""
-        result = self.engine.run_strategy(self.sample_data, self.params)
-        
-        self.assertIsInstance(result, BacktestResult)
-        self.assertIsInstance(result.total_return, float)
-        self.assertIsInstance(result.sharpe_ratio, float)
-        self.assertGreaterEqual(len(result.equity_curve), len(self.sample_data))
-    
-    def test_parameter_validation(self):
-        """Test parameter validation."""
-        with self.assertRaises(ValueError):
-            TradingParameters(rsi_period=0)  # Invalid RSI period
-        
-        with self.assertRaises(ValueError):
-            TradingParameters(ema_fast_period=26, ema_slow_period=12)  # Fast >= Slow
-    
-    def test_empty_data_handling(self):
-        """Test handling of empty data."""
-        empty_data = pd.DataFrame()
-        
-        with self.assertRaises(DataValidationError):
-            self.engine.run_strategy(empty_data, self.params)
-    
-    def test_performance_metrics(self):
-        """Test performance metrics calculation."""
-        result = self.engine.run_strategy(self.sample_data, self.params)
-        
-        # Test that all expected metrics are present and valid
-        self.assertIsInstance(result.sharpe_ratio, float)
-        self.assertIsInstance(result.sortino_ratio, float)
-        self.assertIsInstance(result.max_drawdown, float)
-        self.assertLessEqual(result.max_drawdown, 0)  # Drawdown should be negative
-        self.assertGreaterEqual(result.win_rate, 0)
-        self.assertLessEqual(result.win_rate, 1)
-
-# =================== Advanced Features ===================
-
-class PortfolioManager:
-    """Enhanced portfolio management with advanced risk controls."""
-    
-    def __init__(self, engine: TradingEngine):
-        self._engine = engine
-        self._logger = engine._logger
-        self._risk_budget = {}
-    
-    def run_portfolio_backtest(self, 
-                              assets_data: Dict[str, pd.DataFrame],
-                              params: Dict[str, TradingParameters],
-                              weights: Dict[str, float],
-                              rebalance_freq: str = 'M',
-                              total_capital: float = 100000.0) -> Dict[str, Any]:
-        """
-        Run comprehensive portfolio backtest across multiple assets.
-        
-        Args:
-            assets_data: Dictionary of asset DataFrames
-            params: Dictionary of trading parameters per asset
-            weights: Portfolio weights
-            rebalance_freq: Rebalancing frequency
-            total_capital: Total portfolio capital
-            
-        Returns:
-            Portfolio backtest results
-        """
-        results = {}
-        equity_curves = {}
-        
-        # Validate weights
-        if abs(sum(weights.values()) - 1.0) > 0.001:
-            raise ValueError("Portfolio weights must sum to 1.0")
-        
-        # Run individual asset backtests
-        for asset, data in assets_data.items():
-            if asset not in weights:
-                continue
-            
-            try:
-                asset_params = params.get(asset, TradingParameters())
-                capital_allocation = total_capital * weights[asset]
-                
-                result = self._engine.run_strategy(data, asset_params, capital_allocation)
-                results[asset] = result
-                equity_curves[asset] = result.equity_curve
-                
-                self._logger.info(f"Completed backtest for {asset}: {result.total_return:.2%} return")
-                
-            except Exception as e:
-                self._logger.error(f"Failed to backtest {asset}: {e}")
-                results[asset] = None
-        
-        # Calculate portfolio metrics
-        portfolio_metrics = self._calculate_portfolio_metrics(equity_curves, weights, total_capital)
-        
-        return {
-            'individual_results': results,
-            'portfolio_metrics': portfolio_metrics,
-            'portfolio_equity_curve': portfolio_metrics.get('equity_curve'),
-            'correlation_matrix': self._calculate_correlation_matrix(equity_curves)
-        }
-    
-    def _calculate_portfolio_metrics(self, equity_curves: Dict[str, pd.Series], 
-                                   weights: Dict[str, float], total_capital: float) -> Dict[str, Any]:
-        """Calculate comprehensive portfolio-level metrics."""
-        if not equity_curves:
-            return {}
-        
-        # Align all equity curves to the same time index
-        aligned_curves = pd.DataFrame(equity_curves).fillna(method='ffill').fillna(method='bfill')
-        
-        # Calculate portfolio equity curve
-        portfolio_equity = pd.Series(0.0, index=aligned_curves.index)
-        for asset, curve in aligned_curves.items():
-            if asset in weights:
-                portfolio_equity += curve * (weights[asset] / curve.iloc[0])
-        
-        # Normalize to total capital
-        portfolio_equity = portfolio_equity * total_capital / portfolio_equity.iloc[0]
-        
-        # Calculate portfolio returns
-        portfolio_returns = portfolio_equity.pct_change().dropna()
-        
-        # Portfolio metrics
-        total_return = (portfolio_equity.iloc[-1] / total_capital) - 1.0
-        volatility = portfolio_returns.std() * np.sqrt(252)
-        sharpe_ratio = (portfolio_returns.mean() / portfolio_returns.std()) * np.sqrt(252) if portfolio_returns.std() > 0 else 0.0
-        
-        # Portfolio drawdown
-        peak = portfolio_equity.expanding().max()
-        drawdown = (portfolio_equity - peak) / peak
-        max_drawdown = drawdown.min()
-        
-        return {
-            'equity_curve': portfolio_equity,
-            'total_return': total_return,
-            'annualized_return': (1 + total_return) ** (252 / len(portfolio_equity)) - 1.0,
-            'volatility': volatility,
-            'sharpe_ratio': sharpe_ratio,
-            'max_drawdown': max_drawdown,
-            'calmar_ratio': ((1 + total_return) ** (252 / len(portfolio_equity)) - 1.0) / abs(max_drawdown) if max_drawdown < 0 else 0.0
-        }
-    
-    def _calculate_correlation_matrix(self, equity_curves: Dict[str, pd.Series]) -> pd.DataFrame:
-        """Calculate correlation matrix of asset returns."""
-        if len(equity_curves) < 2:
-            return pd.DataFrame()
-        
-        returns_df = pd.DataFrame()
-        for asset, curve in equity_curves.items():
-            returns_df[asset] = curve.pct_change()
-        
-        return returns_df.corr()
-    
-    def optimize_portfolio_weights(self, assets_data: Dict[str, pd.DataFrame], 
-                                 params: Dict[str, TradingParameters],
-                                 objective: str = 'sharpe_ratio') -> Dict[str, float]:
-        """Optimize portfolio weights using mean-variance optimization."""
-        # This is a simplified implementation
-        # In practice, you'd use scipy.optimize or similar
-        
-        n_assets = len(assets_data)
-        if n_assets < 2:
-            return {list(assets_data.keys())[0]: 1.0} if assets_data else {}
-        
-        # Equal weight as starting point
-        equal_weight = 1.0 / n_assets
-        initial_weights = {asset: equal_weight for asset in assets_data.keys()}
-        
-        # For now, return equal weights
-        # TODO: Implement proper mean-variance optimization
-        return initial_weights
-
-# =================== Risk Analytics Service ===================
-
-class RiskAnalyticsService:
-    """Advanced risk analytics and reporting."""
-    
-    def __init__(self, logger: logging.Logger):
-        self._logger = logger
-    
-    def calculate_risk_metrics(self, equity_curve: pd.Series, benchmark: Optional[pd.Series] = None) -> Dict[str, float]:
-        """Calculate comprehensive risk metrics."""
-        if equity_curve.empty:
-            return {}
-        
-        returns = equity_curve.pct_change().dropna()
-        
-        metrics = {
-            'var_95': self._calculate_var(returns, 0.05),
-            'var_99': self._calculate_var(returns, 0.01),
-            'expected_shortfall_95': self._calculate_expected_shortfall(returns, 0.05),
-            'expected_shortfall_99': self._calculate_expected_shortfall(returns, 0.01),
-            'volatility': returns.std() * np.sqrt(252),
-            'skewness': returns.skew(),
-            'kurtosis': returns.kurtosis(),
-            'downside_deviation': self._calculate_downside_deviation(returns),
-            'ulcer_index': self._calculate_ulcer_index(equity_curve),
-            'pain_index': self._calculate_pain_index(equity_curve)
-        }
-        
-        if benchmark is not None:
-            benchmark_returns = benchmark.pct_change().dropna()
-            metrics.update(self._calculate_benchmark_metrics(returns, benchmark_returns))
-        
-        return metrics
-    
-    def _calculate_var(self, returns: pd.Series, confidence: float = 0.05) -> float:
-        """Calculate Value at Risk."""
-        if returns.empty:
-            return 0.0
-        return returns.quantile(confidence)
-    
-    def _calculate_expected_shortfall(self, returns: pd.Series, confidence: float = 0.05) -> float:
-        """Calculate Expected Shortfall (Conditional VaR)."""
-        if returns.empty:
-            return 0.0
-        var = self._calculate_var(returns, confidence)
-        tail_returns = returns[returns <= var]
-        return tail_returns.mean() if not tail_returns.empty else 0.0
-    
-    def _calculate_downside_deviation(self, returns: pd.Series, target: float = 0.0) -> float:
-        """Calculate downside deviation."""
-        downside_returns = returns[returns < target]
-        return downside_returns.std() * np.sqrt(252) if not downside_returns.empty else 0.0
-    
-    def _calculate_ulcer_index(self, equity_curve: pd.Series) -> float:
-        """Calculate Ulcer Index."""
-        peak = equity_curve.expanding().max()
-        drawdown = ((equity_curve - peak) / peak) * 100
-        drawdown_squared = drawdown ** 2
-        ulcer_index = np.sqrt(drawdown_squared.mean())
-        return ulcer_index
-    
-    def _calculate_pain_index(self, equity_curve: pd.Series) -> float:
-        """Calculate Pain Index."""
-        peak = equity_curve.expanding().max()
-        drawdown = ((equity_curve - peak) / peak) * 100
-        pain_index = abs(drawdown.mean())
-        return pain_index
-    
-    def _calculate_benchmark_metrics(self, returns: pd.Series, benchmark_returns: pd.Series) -> Dict[str, float]:
-        """Calculate benchmark-relative metrics."""
-        # Align series
-        aligned_data = pd.concat([returns, benchmark_returns], axis=1, join='inner')
-        if aligned_data.empty:
-            return {}
-        
-        strategy_returns = aligned_data.iloc[:, 0]
-        bench_returns = aligned_data.iloc[:, 1]
-        
-        # Active returns
-        active_returns = strategy_returns - bench_returns
-        
-        return {
-            'beta': strategy_returns.cov(bench_returns) / bench_returns.var() if bench_returns.var() > 0 else 0.0,
-            'alpha': (strategy_returns.mean() - bench_returns.mean()) * 252,
-            'information_ratio': active_returns.mean() / active_returns.std() * np.sqrt(252) if active_returns.std() > 0 else 0.0,
-            'tracking_error': active_returns.std() * np.sqrt(252),
-            'up_capture': strategy_returns[bench_returns > 0].mean() / bench_returns[bench_returns > 0].mean() if (bench_returns > 0).any() else 0.0,
-            'down_capture': strategy_returns[bench_returns < 0].mean() / bench_returns[bench_returns < 0].mean() if (bench_returns < 0).any() else 0.0
-        }
-
-# =================== Strategy Tester ===================
-
-class StrategyTester:
-    """Comprehensive testing framework for trading strategies."""
-    
-    def __init__(self, engine: TradingEngine):
-        self._engine = engine
-        self._logger = engine._logger
-    
-    def walk_forward_analysis(self, data: pd.DataFrame, 
-                            params: TradingParameters,
-                            train_ratio: float = 0.7,
-                            step_ratio: float = 0.1,
-                            min_train_periods: int = 252) -> pd.DataFrame:
-        """
-        Perform walk-forward analysis with enhanced metrics.
-        
-        Args:
-            data: Historical data
-            params: Trading parameters
-            train_ratio: Ratio of data for training
-            step_ratio: Ratio of data for each step
-            min_train_periods: Minimum training periods
-            
-        Returns:
-            DataFrame with walk-forward results
-        """
-        results = []
-        data_length = len(data)
-        
-        train_size = max(int(data_length * train_ratio), min_train_periods)
-        step_size = max(int(data_length * step_ratio), 30)
-        
-        for start_idx in range(0, data_length - train_size - step_size, step_size):
-            train_end = start_idx + train_size
-            test_end = min(train_end + step_size, data_length)
-            
-            if test_end - train_end < 30:  # Ensure minimum test period
-                continue
-            
-            train_data = data.iloc[start_idx:train_end]
-            test_data = data.iloc[train_end:test_end]
-            
-            try:
-                # In practice, optimize parameters on train_data
-                # For now, use provided parameters
-                result = self._engine.run_strategy(test_data, params)
-                
-                results.append({
-                    'period_start': test_data.index[0],
-                    'period_end': test_data.index[-1],
-                    'train_periods': len(train_data),
-                    'test_periods': len(test_data),
-                    'total_return': result.total_return,
-                    'annualized_return': result.annualized_return,
-                    'sharpe_ratio': result.sharpe_ratio,
-                    'sortino_ratio': result.sortino_ratio,
-                    'max_drawdown': result.max_drawdown,
-                    'volatility': result.volatility,
-                    'total_trades': result.total_trades,
-                    'win_rate': result.win_rate,
-                    'profit_factor': result.profit_factor
-                })
-                
-            except Exception as e:
-                self._logger.error(f"Walk-forward test failed for period {start_idx}: {e}")
-        
-        return pd.DataFrame(results)
-    
-    def monte_carlo_simulation(self, data: pd.DataFrame, 
-                              params: TradingParameters,
-                              n_simulations: int = 1000,
-                              block_length: int = 30) -> Dict[str, np.ndarray]:
-        """
-        Run Monte Carlo simulations with block bootstrap.
-        
-        Args:
-            data: Historical data
-            params: Trading parameters
-            n_simulations: Number of simulations
-            block_length: Block length for bootstrap
-            
-        Returns:
-            Dictionary of simulation results
-        """
-        results = {
-            'returns': [], 
-            'sharpe_ratios': [], 
-            'max_drawdowns': [],
-            'volatilities': [],
-            'win_rates': [],
-            'total_trades': []
-        }
-        
-        # Calculate returns for block bootstrap
-        returns = data['close'].pct_change().dropna()
-        
-        for i in range(n_simulations):
-            try:
-                # Generate synthetic data using block bootstrap
-                synthetic_returns = self._block_bootstrap(returns, len(data), block_length)
-                synthetic_data = self._returns_to_ohlcv(synthetic_returns, data.iloc[0])
-                
-                result = self._engine.run_strategy(synthetic_data, params)
-                
-                results['returns'].append(result.total_return)
-                results['sharpe_ratios'].append(result.sharpe_ratio)
-                results['max_drawdowns'].append(result.max_drawdown)
-                results['volatilities'].append(result.volatility)
-                results['win_rates'].append(result.win_rate)
-                results['total_trades'].append(result.total_trades)
-                
-            except Exception as e:
-                self._logger.warning(f"Monte Carlo simulation {i} failed: {e}")
-        
-        return {k: np.array(v) for k, v in results.items()}
-    
-    def _block_bootstrap(self, returns: pd.Series, target_length: int, block_length: int) -> pd.Series:
-        """Perform block bootstrap on returns."""
-        n_blocks = int(np.ceil(target_length / block_length))
-        bootstrapped_returns = []
-        
-        for _ in range(n_blocks):
-            start_idx = np.random.randint(0, len(returns) - block_length + 1)
-            block = returns.iloc[start_idx:start_idx + block_length]
-            bootstrapped_returns.extend(block.values)
-        
-        # Trim to target length
-        bootstrapped_returns = bootstrapped_returns[:target_length]
-        
-        return pd.Series(bootstrapped_returns)
-    
-    def _returns_to_ohlcv(self, returns: pd.Series, initial_data: pd.Series) -> pd.DataFrame:
-        """Convert returns to OHLCV format."""
-        prices = initial_data['close'] * (1 + returns).cumprod()
-        
-        # Simple OHLCV generation
-        data = pd.DataFrame(index=pd.date_range(start='2020-01-01', periods=len(prices), freq='D'))
-        data['close'] = prices.values
-        data['open'] = data['close'].shift(1).fillna(data['close'].iloc[0])
-        
-        # Generate high/low with some noise
-        noise = np.random.normal(0, 0.002, len(prices))
-        data['high'] = data['close'] * (1 + abs(noise))
-        data['low'] = data['close'] * (1 - abs(noise))
-        data['volume'] = np.random.randint(1000, 10000, len(prices))
-        
-        return data
-
-# =================== Factory and Configuration ===================
-
-class TradingEngineFactory:
-    """Enhanced factory for creating pre-configured trading engines."""
-    
-    @staticmethod
-    def create_default_engine(config: Optional[EngineConfig] = None) -> TradingEngine:
-        """Create engine with default configuration."""
-        return TradingEngine(config=config or EngineConfig())
-    
-    @staticmethod
-    def create_conservative_engine() -> TradingEngine:
-        """Create conservative engine with low-risk settings."""
-        config = EngineConfig(
-            max_position_size=0.5,
-            max_portfolio_risk=0.01,
-            max_drawdown_threshold=0.10,
-            volatility_threshold=0.20
-        )
-        return TradingEngine(config=config)
-    
-    @staticmethod
-    def create_aggressive_engine() -> TradingEngine:
-        """Create aggressive engine with higher risk tolerance."""
-        config = EngineConfig(
-            max_position_size=1.5,
-            max_portfolio_risk=0.05,
-            max_drawdown_threshold=0.25,
-            volatility_threshold=0.40
-        )
-        return TradingEngine(config=config)
-    
-    @staticmethod
-    def create_test_engine() -> TradingEngine:
-        """Create engine optimized for testing."""
-        config = EngineConfig(
-            cache_indicators=False,
-            log_level='ERROR',
-            min_data_points=50
-        )
-        return TradingEngine(config=config)
-
-# =================== Enhanced Usage Example ===================
-
-def enhanced_example_usage():
-    """Enhanced example demonstrating all features."""
-    
-    # Create sample data with more realistic characteristics
-    dates = pd.date_range(start='2020-01-01', end='2023-12-31', freq='D')
-    np.random.seed(42)
-    
-    # Generate more realistic price series with trends and volatility clusters
-    n_points = len(dates)
-    base_returns = np.random.normal(0.0005, 0.02, n_points)
-    
-    # Add trend components
-    trend_component = np.sin(np.linspace(0, 4*np.pi, n_points)) * 0.001
-    base_returns += trend_component
-    
-    # Add volatility clustering
-    volatility = np.random.exponential(0.01, n_points)
-    returns = base_returns * volatility
-    
-    prices = 100 * (1 + returns).cumprod()
-    
-    # Create realistic OHLCV data
-    sample_data = pd.DataFrame({
-        'open': prices * (1 + np.random.normal(0, 0.001, n_points)),
-        'close': prices,
-        'high': prices * (1 + abs(np.random.normal(0, 0.005, n_points))),
-        'low': prices * (1 - abs(np.random.normal(0, 0.005, n_points))),
-        'volume': np.random.lognormal(8, 0.5, n_points).astype(int)
-    }, index=dates)
-    
-    # Ensure OHLC relationships
-    sample_data['high'] = np.maximum.reduce([
-        sample_data['open'], sample_data['high'], 
-        sample_data['low'], sample_data['close']
-    ])
-    sample_data['low'] = np.minimum.reduce([
-        sample_data['open'], sample_data['high'], 
-        sample_data['low'], sample_data['close']
-    ])
-    
-    # Create enhanced trading engine
-    config = EngineConfig(
-        enable_stop_loss=True,
-        enable_take_profit=True,
-        cache_indicators=True,
-        log_level='INFO'
-    )
-    
-    engine = TradingEngineFactory.create_default_engine(config)
-    
-    # Enhanced parameters with better defaults
-    params = TradingParameters(
-        rsi_period=14,
-        rsi_oversold=25,
-        rsi_overbought=75,
-        ema_fast_period=12,
-        ema_slow_period=26,
-        signal_threshold=0.15,
-        stop_loss_atr_mult=1.5,
-        take_profit_atr_mult=2.5,
-        ensemble_weights={
-            'trend': 0.35,
-            'mean_reversion': 0.15,
-            'momentum': 0.35,
-            'volatility_breakout': 0.15
-        }
-    )
-    
-    print("=" * 80)
-    print("ENHANCED TRADING STRATEGY BACKTEST RESULTS")
-    print("=" * 80)
-    
-    # Run comprehensive backtest
-    result = engine.run_strategy(sample_data, params, initial_capital=100000, fee_bps=5)
-    
-    # Display comprehensive results
-    print(f"\nPERFORMANCE METRICS:")
-    print(f"{'Total Return:':<25} {result.total_return:>8.2%}")
-    print(f"{'Annualized Return:':<25} {result.annualized_return:>8.2%}")
-    print(f"{'Volatility:':<25} {result.volatility:>8.2%}")
-    print(f"{'Sharpe Ratio:':<25} {result.sharpe_ratio:>8.2f}")
-    print(f"{'Sortino Ratio:':<25} {result.sortino_ratio:>8.2f}")
-    print(f"{'Calmar Ratio:':<25} {result.calmar_ratio:>8.2f}")
-    print(f"{'Omega Ratio:':<25} {result.omega_ratio:>8.2f}")
-    
-    print(f"\nRISK METRICS:")
-    print(f"{'Max Drawdown:':<25} {result.max_drawdown:>8.2%}")
-    print(f"{'VaR (95%):':<25} {result.var_95:>8.2%}")
-    print(f"{'Expected Shortfall:':<25} {result.expected_shortfall_95:>8.2%}")
-    print(f"{'Tail Ratio:':<25} {result.tail_ratio:>8.2f}")
-    
-    print(f"\nTRADING METRICS:")
-    print(f"{'Total Trades:':<25} {result.total_trades:>8d}")
-    print(f"{'Win Rate:':<25} {result.win_rate:>8.2%}")
-    print(f"{'Profit Factor:':<25} {result.profit_factor:>8.2f}")
-    print(f"{'Avg Trade Duration:':<25} {str(result.avg_trade_duration).split(',')[0]:>8s}")
-    print(f"{'Largest Win:':<25} ${result.largest_win:>7,.2f}")
-    print(f"{'Largest Loss:':<25} ${result.largest_loss:>7,.2f}")
-    
-    # Check for alerts
-    alerts = engine.get_performance_alerts()
-    if alerts:
-        print(f"\nPERFORMANCE ALERTS:")
-        for alert in alerts:
-            print(f"  • {alert}")
-    
-    # Run parameter optimization
-    print(f"\nPARAMETER OPTIMIZATION:")
-    param_ranges = {
-        'rsi_period': [10, 14, 20],
-        'ema_fast_period': [8, 12, 16],
-        'ema_slow_period': [21, 26, 30],
-        'signal_threshold': [0.1, 0.15, 0.2]
-    }
-    
-    best_params, best_score = engine.optimize_parameters(
-        sample_data, param_ranges, objective='sharpe_ratio', max_iterations=50
-    )
-    
-    print(f"Best Parameters Found:")
-    print(f"  RSI Period: {best_params.rsi_period}")
-    print(f"  EMA Fast: {best_params.ema_fast_period}")
-    print(f"  EMA Slow: {best_params.ema_slow_period}")
-    print(f"  Signal Threshold: {best_params.signal_threshold}")
-    print(f"  Best Sharpe Ratio: {best_score:.3f}")
-    
-    # Demonstrate portfolio management
-    print(f"\nPORTFOLIO SIMULATION:")
-    portfolio_manager = PortfolioManager(engine)
-    
-    # Simulate multiple assets
-    assets_data = {
-        'ASSET_A': sample_data,
-        'ASSET_B': sample_data * 1.1,  # Slightly different asset
-        'ASSET_C': sample_data * 0.9   # Another variation
-    }
-    
-    portfolio_params = {
-        'ASSET_A': params,
-        'ASSET_B': params,
-        'ASSET_C': params
-    }
-    
-    portfolio_weights = {'ASSET_A': 0.4, 'ASSET_B': 0.35, 'ASSET_C': 0.25}
-    
-    portfolio_result = portfolio_manager.run_portfolio_backtest(
-        assets_data, portfolio_params, portfolio_weights, total_capital=100000
-    )
-    
-    portfolio_metrics = portfolio_result['portfolio_metrics']
-    print(f"Portfolio Return: {portfolio_metrics['total_return']:.2%}")
-    print(f"Portfolio Sharpe: {portfolio_metrics['sharpe_ratio']:.2f}")
-    print(f"Portfolio Max DD: {portfolio_metrics['max_drawdown']:.2%}")
-    
-    print("=" * 80)
-    print("BACKTEST COMPLETED SUCCESSFULLY")
-    print("=" * 80)
-    
-    return result, best_params, portfolio_result
-
-# =================== Export Interface ===================
-
-__all__ = [
-    'TradingEngine',
-    'TradingEngineFactory',
-    'TradingParameters',
-    'EngineConfig',
-    'BacktestResult',
-    'TechnicalIndicators',
-    'Trade',
-    'SignalType',
-    'MarketRegime',
-    'OrderType',
-    'PortfolioManager',
-    'RiskAnalyticsService',
-    'PerformanceMonitor',
-    'StrategyTester',
-    'TestTradingEngine',
-    'enhanced_example_usage',
-    # Exceptions
-    'TradingEngineError',
-    'DataValidationError',
-    'IndicatorComputationError',
-    'BacktestExecutionError',
-    'InsufficientDataError',
-    'ConfigurationError',
-    'RiskLimitExceededError'
-]
-
-# === Backward-compat shim for TradingGUI (drop-in, no UI changes) ===
-from dataclasses import dataclass
-from typing import Dict, Tuple, Optional
-import pandas as pd
-import logging
-
-# Używamy istniejących bytów z tego modułu:
-# - TradingEngine, TradingEngineFactory, TradingParameters, TradingSignalService
-
-class _NoShortSignalService(TradingSignalService):
-    """Wariant generatora sygnałów, który usuwa shorty (sygnały < 0 -> 0)."""
-    def generate_signals(self, indicators, params):
-        sig = super().generate_signals(indicators, params)
-        # Wyłącz shorty (zachowanie zgodne z allow_short=False w starym GUI)
-        return sig.where(sig > 0, 0.0)
-
-class TradingStrategies:
-    """
-    Back-compat wrapper oczekiwany przez TradingGUI._run_backtest():
-      backtest(data, initial_capital, fee, slippage, fraction, allow_short)
-      -> (metrics: dict, trades_df: pd.DataFrame, equity: pd.Series)
-
-    Zasada:
-    - Mapujemy stare parametry GUI na TradingParameters (z domyślnymi ustawieniami).
-    - Fee+slippage konwertujemy do bps.
-    - Jeśli allow_short=False, korzystamy z wariantu silnika bez shortów (bez zmian w GUI).
-    """
-
-    def __init__(self, engine: Optional[TradingEngine] = None, logger: Optional[logging.Logger] = None):
-        self._base_engine = engine or TradingEngineFactory.create_default_engine()
-        self._logger = logger or logging.getLogger("TradingStrategiesShim")
-
-    def _mk_engine(self, allow_short: bool) -> TradingEngine:
-        if allow_short:
-            return self._base_engine
-        # Stwórz silnik z generatorem bez shortów; bez ingerencji w GUI.
-        return TradingEngine(
-            config=self._base_engine._config if hasattr(self._base_engine, "_config") else None,
-            validator=None,  # użyj domyślnych
-            indicator_calculator=None,
-            signal_generator=_NoShortSignalService(self._logger),
-            risk_manager=None,
-            backtest_engine=None,
-            logger=self._logger
-        )
-
-    def backtest(
-        self,
-        data: pd.DataFrame,
-        initial_capital: float = 10000.0,
-        fee: float = 0.0004,
-        slippage: float = 0.0002,
-        fraction: float = 0.05,
-        allow_short: bool = False,
-        # future-proof: opcjonalny most AI bez zmiany GUI
-        ai_model=None,
-        ai_weight: float = 0.0,
-        ai_threshold_bps: float = 5.0,
-    ) -> Tuple[Dict, pd.DataFrame, pd.Series]:
-        # 1) Mapowanie GUI -> parametry strategii (bezpieczne domyślne)
-        params = TradingParameters(
-            # Wskaźniki – domyślne wartości nowego silnika
-            rsi_period=14, rsi_oversold=30.0, rsi_overbought=70.0,
-            ema_fast_period=12, ema_slow_period=26, sma_trend_period=50,
-            bb_period=20, bb_std_mult=2.0, atr_period=14, macd_signal_period=9,
-            stoch_k_period=14, stoch_d_period=3,
-            signal_threshold=0.10,
-            # Ryzyko / sizing
-            stop_loss_atr_mult=2.0, take_profit_atr_mult=3.0,
-            position_size=float(max(0.0, min(1.0, fraction))),
-            max_position_risk=0.02,
-            volatility_target=0.15, kelly_fraction=0.25,
-        )
-
-        # 2) Opłaty w bps (fee + slippage)
-        fee_bps = float((fee + slippage) * 10000.0)
-
-        # 3) Wybór silnika z/do shortów zgodnie z allow_short
-        engine = self._mk_engine(bool(allow_short))
-
-        # 4) (Opcjonalnie) fuzja AI+TA – tylko jeśli przekażesz ai_model (GUI nie musi tego robić)
-        #    Bez zmian GUI: brak ai_model => czysta TA.
-        if ai_model is None or ai_weight <= 0.0:
-            result = engine.run_strategy(
-                data=data,
-                params=params,
-                initial_capital=float(initial_capital),
-                fee_bps=fee_bps
-            )
-        else:
-            # Lekki hook: jeśli w przyszłości przekażesz ai_model, zastosuj most (patrz bridges/).
-            try:
-                from bridges.ai_trading_bridge import AITradingBridge
-            except Exception:
-                AITradingBridge = None
-
-            if AITradingBridge is None:
-                result = engine.run_strategy(data=data, params=params, initial_capital=float(initial_capital), fee_bps=fee_bps)
-            else:
-                # Uzyskaj wewnętrzne komponenty, policz wskaźniki i surowe sygnały, wstrzyknij fuzję AI, potem risk+backtest.
-                # (Zachowuje semantykę engine.run_strategy, ale z fuzją.)
-                validated_data = engine._validator.validate_ohlcv(data)  # używa istniejących serwisów
-                indicators = engine._indicator_calculator.calculate_indicators(validated_data, params)
-                raw_signals = engine._signal_generator.generate_signals(indicators, params)
-                bridge = AITradingBridge(ai_model, weight_ai=float(ai_weight), threshold_bps=float(ai_threshold_bps))
-                fused_signals = bridge.integrate(validated_data, raw_signals)
-                managed_signals = engine._risk_manager.apply_risk_management(validated_data, fused_signals, indicators, params)
-                result = engine._backtest_engine.run_backtest(validated_data, managed_signals, params, initial_capital, fee_bps, engine._config)
-
-        # 5) Konwersja wyników do formatu GUI
-        metrics = {
-            "total_return": result.total_return,
-            "annualized_return": result.annualized_return,
-            "volatility": result.volatility,
-            "sharpe_ratio": result.sharpe_ratio,
-            "sortino_ratio": result.sortino_ratio,
-            "calmar_ratio": result.calmar_ratio,
-            "omega_ratio": result.omega_ratio,
-            "max_drawdown": result.max_drawdown,
-            "win_rate": result.win_rate,
-            "profit_factor": result.profit_factor,
-            "trades": int(result.total_trades),
-            # kilka dodatkowych, jeśli GUI je kiedyś wyświetli:
-            "tail_ratio": result.tail_ratio,
-            "var_95": result.var_95,
-            "expected_shortfall_95": result.expected_shortfall_95,
-        }
-        return metrics, result.trades, result.equity_curve
-
-
-# Dopiszemy do __all__ bez ręcznej edycji istniejącej listy
-try:
-    if 'TradingStrategies' not in __all__:
-        __all__.append('TradingStrategies')
-except Exception:
-    __all__ = ['TradingStrategies']
-
-
-if __name__ == "__main__":
-    # Run enhanced example with comprehensive testing
-    try:
-        # Run unit tests first
-        print("Running unit tests...")
-        unittest.main(argv=[''], exit=False, verbosity=2)
-        
-        # Run enhanced example
-        print("\nRunning enhanced example...")
-        result, best_params, portfolio_result = enhanced_example_usage()
-        
-        print(f"\nExample completed successfully!")
-        
-    except Exception as e:
-        print(f"Error in main execution: {e}")
-        raise
+from KryptoLowca.strategies import *  # noqa: F401,F403
